requirements.txt
phidata
streamlit
openai
duckdb
pandas
numpy
agno